django_backend/requirements.txt
Django>=4.2,<5
djangorestframework
gunicorn
psycopg2-binary
django-cors-headers
google-cloud-pubsub>=2.20.0
